Changes to Hero, Method, and Character.
Now there are 4 different heroes to choose from; creation of a hero still is done with a single choice-argument.
Method for exchanging attacks between a mosnter and a hero has been added to the Method-class. The method takes for paramenters Character (which is now basically is a monster; later this will be change to Monster ofc) and Hero objects.
And, lastly, Charcter class lost most of its useless constructors and
object creation is no longer done through setters, but by directly
applying edditing instance variables.