I prefer the unedited version of my code.
I noticed few changes on my code.
I didn't agree with them so I reverted this edit.
I did that manually, in order to avoid to edit my editors work, on his code.

Few thoughts. in the method class. it was a *main method* that Akhmed left as testing script for the others. it was not edited, and that makes me thing that the changes possibly were made directly through the web browser.
And that makes me thinking. *Is it possible that* the one who edited the basic functionality of 20 hours-work and havily tested structure , (one day before the presentation that I have to do in the meeting as tester), did that through the web browser?
I assume that is not the case amd I really hope that is not the case, and I leave it there because the changes were small and revertable  anyway.

I really wish to ask, that if someone does not have my permission explicitely, to not edit my code,

If I see that my code is touched again without my permission I will have to reduce my productivity to the  minimum required code or even less.

My code is important thing for me and extremelly pesronal.