added clearScreen method (Not Implemented!!!)
I added clearScreen method that clears screen in cmd.
To be implemented or deleted if there is no use for it.

My idea was to use it to make our game readable and clean.
Of course, there is another method with 50 empty lines:
for(int i=0; i<50; i++){
sout();
}
Feel free to use or delete :)