Save/Load now is fully functional
bugs can be around the corner but I was unable to detect any bug and it seems that it works nicely.
Code is bit redudant but I decided to leave it that way in order to avoid bugs.
Please be kind to test everything as much as possible and tell me if something is wrong.