Ensured exit gateway
Now the map will always generate at least one exit door.
I also reduced the possibility of facing a wall instead of door, in order to avoid lockdown.
There is still the possibility to hit a wall without a door on your way, but the possibility of lockdown in one or two rooms is reduced.
This system can't be 100% foolproof