the fighting system now works.
The problem came because of incosistency in the design between three classes. Insted of changing everything in some unlucky class or method, I did  a small 5 line patch that solves the problem.
I got forced for practical reasons to  add the patch in the Room class, and more specifically in the main constructor of the Room class